openclaw/requirements.txt
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
